requirements.txt
# Core ARI and WebSocket dependencies
ari==0.1.3
websockets==15.0.1
aiohttp==3.13.3

# Configuration and logging
pydantic>=2.7.0
pydantic_settings==2.12.0
structlog==25.5.0
prometheus-client==0.24.1
PyYAML==6.0.3

# Audio resampling (numpy linear interpolation for artifact-free streaming)
numpy>=1.24.0

# Utilities
tenacity==8.2.3

# WebRTC VAD for robust speech detection
webrtcvad==2.0.10

# Email tools dependencies
resend>=2.0.0
jinja2==3.1.6
dnspython==2.8.0

# NumPy for audio resampling (replaces audioop.ratecv)
numpy>=1.24.0

# OpenAI for summary generation
openai>=1.0.0
